# src/frontend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Fitness Tracking API
  version: 1.0.0
  description: |
    A comprehensive fitness tracking API with voice integration, workout management, 
    exercise libraries, and LLM-powered workout generation capabilities.
  
    ## Features
    - User authentication with JWT tokens
    - Workout creation and tracking
    - Exercise library with media support
    - Voice-powered workout logging
    - LLM workout generation
    - Goal setting and progress tracking
    - Offline support with sync capabilities
    
  contact:
    name: API Support
    email: support@fitnessapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fitnessapp.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /voice/sessions:
    get:
      tags:
        - Voice Integration
      summary: Get voice sessions
      description: Retrieve user's voice sessions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Voice sessions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/VoiceSession'

    post:
      tags:
        - Voice Integration
      summary: Create voice session
      description: Create a new voice session for workout logging
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                workoutId:
                  type: string
                  description: Optional workout ID to associate with session
      responses:
        '201':
          description: Voice session created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VoiceSession'
  /voice/sessions/{id}:
    get:
      tags:
        - Voice Integration
      summary: Get voice session
      description: Retrieve a specific voice session
      parameters:
        - $ref: '#/components/parameters/VoiceSessionId'
      responses:
        '200':
          description: Voice session retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VoiceSession'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
...
    VoiceSessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Voice session ID
      example: "voice_session_101"
  schemas:
...
    VoiceSession:
      type: object
      required:
        - id
        - userId
        - startTime
        - transcript
        - processed
        - status
      properties:
        id:
          type: string
          example: "voice_session_123"
        userId:
          type: string
          example: "user_123"
        workoutId:
          type: string
          example: "workout_123"
        startTime:
          type: string
          format: date-time
          example: "2024-01-15T09:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-01-15T09:05:00Z"
        transcript:
          type: string
          example: "I just completed 3 sets of bench press with 135 pounds for 8 reps each"
        processed:
          type: boolean
          example: true
        extractedData:
          type: object
          description: Structured data extracted from voice
          example:
            exercises: ["bench_press"]
            sets: [{"reps": 8, "weight": 135}]
        status:
          $ref: '#/components/schemas/VoiceSessionStatus'

    # Enums
...
    VoiceSessionStatus:
      type: string
      enum:
        - recording
        - processing
        - completed
        - error
      example: completed

    # Request Schemas
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
...